requirements.txt
asgiref==3.6.0
attrs==23.1.0
Django==4.2.1
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
drf-spectacular-sidecar==2023.5.1
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.17.3
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
uritemplate==4.1.1
zope.event==4.6
zope.interface==6.0
